backend/requirements.txt
django
django-cors-headers
djangorestframework
openrouteservice
folium